\name{bootsKey}
\alias{bootsKey}
\title{Time Series Key Object}
\description{
  A time series keys for the \code{\link{boots}} time
  series format is an S4 object itself.
}
\section{Slots}{
  \describe{ \item{\code{country}:}{Object of class
  \code{"character"}} \item{\code{provider}:}{Object of
  class \code{"character"}} \item{\code{src}:}{Object of
  class \code{"character"}} \item{\code{level}:}{Object of
  class \code{"character"}}
  \item{\code{selectedLevel}:}{Object of class
  \code{"character"}} \item{\code{variable}:}{Object of
  class \code{"character"}} \item{\code{item}:}{Object of
  class \code{"character"}} }
}
\author{
  Matthias Bannert
}
\seealso{
  \code{\link{boots}}
}

