\docType{methods}
\name{getMeta}
\alias{getMeta}
\title{Show Localized Meta Information}
\arguments{
  \item{object}{an R object of class
  \code{\link{metalocalized}},}

  \item{lang}{character string denoting the language. Only
  use available abbrebvation displayed when calling the
  object.}
}
\value{
  list containing key value pairs.
}
\description{
  This methods displays meta information stored in an
  object of class \code{\link{metalocalized}}. Its return
  is basically a list. It also returns the abbreviation of
  available languages.
}
\author{
  Matthias Bannert
}
\seealso{
  \code{\link{boots}} and \code{\link{metalocalized}}
}

