\name{is.metalocalized}
\alias{is.metalocalized}
\title{Test for a metalocalized object}
\usage{
  is.metalocalized(x)
}
\arguments{
  \item{x}{object whose class should be checked}
}
\description{
  This function tests whether an object is of class
  metalocalized. It's rather not used directly.
}
\author{
  Matthias Bannert
}

