\name{is.tsDbResult}
\alias{is.tsDbResult}
\title{Test for an object of class tsDbResult}
\usage{
  is.tsDbResult(x)
}
\arguments{
  \item{x}{object whose class should be checked}
}
\description{
  This function tests whether an object is of class
  tsDbResult. It's rather not used directly.
}
\author{
  Matthias Bannert
}

