\name{metalocalized}
\alias{metalocalized}
\title{Class to Store Localized Meta Information}
\description{
  This class stores localized meta information in a nested
  list. Each sublist refers to one particular language.
  This class is rather not used directly but as a part of
  the \code{\link{boots}} class.
}
\author{
  Matthias Bannert
}

