\docType{methods}
\name{appendMeta}
\alias{appendMeta}
\alias{replaceMeta}
\title{Extend or Replace Localized Meta Information}
\arguments{
  \item{object}{an R object of class
  \code{\link{metalocalized}} or \code{\link{boots}}.}
}
\description{
  The methods appendMeta and replaceMeta replace replace or
  append localized meta information stored objects of class
  \code{\link{metalocalized}}.
}
\author{
  Matthias Bannert
}
\seealso{
  \code{\link{boots}} and \code{\link{metalocalized}}
}

