\docType{methods}
\name{plot}
\alias{plot}
\title{Simple plot method for boots}
\arguments{
  \item{x}{object of class \code{\link{boots}}}

  \item{...}{arguments that are passed to the standard plot
  function}

  \item{l}{character string indicating the language of the
  tile.}
}
\description{
  This is a preliminary plotting method for objects of
  class boots. Basically turns boots objects to standard ts
  objects and adds meta information. Currently only
  quarterly data is fully supported.
}
\author{
  Matthias Bannert
}

