\name{tsDbResult}
\alias{tsDbResult}
\title{Class to Store Query Results Temporarily}
\description{
  This helper class stores query results from queries to
  the Postgres database temporarily. It is only used to
  induce the corresponding method when initializing new
  objects of class \code{\link{boots}}.
}
\author{
  Matthias Bannert
}

